+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [UCX-3639] Fixed display order of mentions in the flex panel
 * [UCX-3638] Fixed issue with Bot responses displayed multiple times
 * [UCX-3640] Fixed Bot avatar on page reload
+* [UCX-3635] Fixed user able to delete their own message
 
 ### Bug Fixes